코드 정리

+++ GDIGameTemplate/RenderSystem.cpp
@@ -155,7 +155,7 @@ namespace Render
 		Gdiplus::Rect destRect(x, y,  srcRect.Width, srcRect.Height); // 화면에 그릴 영역
 		if (mirror)	//	Y 축 대칭 인가?
 		{	
-			static Gdiplus::Matrix matrixMirror(-1, 0, 0, 1, srcRect.Width,0);
+			static Gdiplus::Matrix matrixMirror(-1, 0, 0, 1,(float) srcRect.Width,0);
 			graphics->SetTransform(&matrixMirror);
 		}
 		else
